13-03-2024

+++ ProjetPI(back)/src/main/java/tn/esprit/backend/Entite/Rating.java
@@ -16,6 +16,6 @@ public class Rating {
     private int idUser;
     private int idFeedback;
     private String status;
-    private double moyrating;
+    private int moyrating;
 
 }
